MCR-2685 upload file name check with regexp (#1604)
