+++ mycore-webtools/src/main/resources/components/webtools/config/mycore.properties
@@ -5,3 +5,5 @@
 
 MCR.Jersey.Resource.Packages=%MCR.Jersey.Resource.Packages%,org.mycore.webtools.session,org.mycore.webtools.upload
 MCR.URIResolver.xslIncludes.components=%MCR.URIResolver.xslIncludes.components%,sessionListing.xsl,processingGUI.xsl,msie-warning.xsl
+
+MCR.FileUpload.FileNamePattern=\\A.+\\z
