+++ mycore-webtools/src/main/ts/processing/package.json
