+++ mycore-webtools/src/main/ts/upload/package.json
@@ -1,0 +1,11 @@
+{
+  "scripts": {
+    "build": "webpack --progress"
+  },
+  "devDependencies": {
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.7.4",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.10.0"
+  }
+}
